# packages/contracts/src/openapi.yaml
openapi: 3.1.0
info:
  title: Xenitex Appliance API
  version: 1.0.0
  description: >
    v1 of the on-premise appliance API. This document is the single source of
    truth (P1-01) — the TypeScript client and types in this package are
    generated from it, never hand-written (P1-02).

    Conventions (see docs/adr/0006-api-conventions.md, accepted at GATE 1):
    opaque cursor pagination on every collection (`?cursor=&limit=`, response
    `{ items, nextCursor }`); RFC 9457 Problem Details on every non-2xx
    response; `Idempotency-Key` required on every creation POST;
    `ETag`/`If-Match` required on every mutation of a mutable resource; URI
    path versioning (`/v1/...`); long-running operations return `202` with a
    `Location` header pointing at a pollable resource, never a blocking
    request.

    JSON casing: camelCase, matching packages/domain. Any field whose value
    originates from a scanned target (SEC-17/AIP-02 — banners, headers,
    certificate subjects, hostnames, identity-key values, OS/service/product
    inference) carries an `Untrusted` suffix in its property name. This is
    the same greppable-naming convention as the `_untrusted` column suffix in
    docs/adr/0007-schema-conventions.md, extended to the API boundary so a
    frontend consumer can never mistake such a value for first-party data —
    it must be escaped on output, never rendered as HTML, never interpreted.
servers:
  - url: /v1
...
security:
  - sessionCookie: []
paths:
...
  /authorized-scopes:
    get:
      tags: [Scopes]
      operationId: listAuthorizedScopes
      summary: List authorized scopes
      parameters:
        [{ $ref: '#/components/parameters/Cursor' }, { $ref: '#/components/parameters/Limit' }]
      responses:
        '200':
          description: 'OK.'
          content:
            application/json:
              schema:
                type: object
                required: [items, nextCursor]
                properties:
                  items: { type: array, items: { $ref: '#/components/schemas/AuthorizedScope' } }
                  nextCursor: { oneOf: [{ type: string }, { type: 'null' }] }
    post:
      tags: [Scopes]
      operationId: createAuthorizedScope
      summary: Declare an authorized scope (SAFE-01). Immutable once accepted.
      parameters: [{ $ref: '#/components/parameters/IdempotencyKey' }]
      requestBody:
        required: true
        content:
          { application/json: { schema: { $ref: '#/components/schemas/AuthorizedScopeCreate' } } }
      responses:
        '201':
          {
            description: 'Created.',
            content:
              { application/json: { schema: { $ref: '#/components/schemas/AuthorizedScope' } } },
          }
        '400': { $ref: '#/components/responses/ValidationError' }
  /authorized-scopes/{scopeId}:
    parameters:
      - name: scopeId
        in: path
        required: true
        schema: { type: string }
    get:
      tags: [Scopes]
      operationId: getAuthorizedScope
      summary: Get an authorized scope
      responses:
        '200':
          {
            description: 'OK.',
            content:
              { application/json: { schema: { $ref: '#/components/schemas/AuthorizedScope' } } },
          }
        '404': { $ref: '#/components/responses/NotFound' }
  /authorized-scopes/{scopeId}/supersede:
    parameters:
      - name: scopeId
        in: path
        required: true
        schema: { type: string }
    post:
      tags: [Scopes]
      operationId: supersedeAuthorizedScope
      summary: >-
        SAFE-01: scopes are immutable once accepted, so "editing" one means
        declaring a brand-new scope (its own fresh attestation) that
        supersedes this one. The old record's fields never change — only its
        supersededById link is set, exactly once.
      parameters: [{ $ref: '#/components/parameters/IdempotencyKey' }]
      requestBody:
        required: true
        content:
          { application/json: { schema: { $ref: '#/components/schemas/AuthorizedScopeCreate' } } }
      responses:
        '201':
          {
            description: 'Created — the new, superseding scope.',
            content:
              { application/json: { schema: { $ref: '#/components/schemas/AuthorizedScope' } } },
          }
        '400': { $ref: '#/components/responses/ValidationError' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409':
          description: This scope has already been superseded by a later declaration.
          content:
            {
              application/problem+json: { schema: { $ref: '#/components/schemas/ProblemDetails' } },
            }

  # ------------------------------------------------------------ ExclusionRules
  /exclusion-rules:
    get:
      tags: [ExclusionRules]
      operationId: listExclusionRules
      summary: List exclusion rules
      parameters:
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
        - name: scopeId
          in: query
          schema: { type: string }
      responses:
        '200':
          description: 'OK.'
          content:
            application/json:
              schema:
                type: object
                required: [items, nextCursor]
                properties:
                  items: { type: array, items: { $ref: '#/components/schemas/ExclusionRule' } }
                  nextCursor: { oneOf: [{ type: string }, { type: 'null' }] }
    post:
      tags: [ExclusionRules]
      operationId: createExclusionRule
      summary: SAFE-02 — enforced at plan validation and again in the worker before dispatch
      parameters: [{ $ref: '#/components/parameters/IdempotencyKey' }]
      requestBody:
        required: true
        content:
          { application/json: { schema: { $ref: '#/components/schemas/ExclusionRuleCreate' } } }
      responses:
        '201':
          {
            description: 'Created.',
            content:
              { application/json: { schema: { $ref: '#/components/schemas/ExclusionRule' } } },
          }
        '400': { $ref: '#/components/responses/ValidationError' }
  /exclusion-rules/{ruleId}:
    parameters:
      - name: ruleId
        in: path
        required: true
        schema: { type: string }
    get:
      tags: [ExclusionRules]
      operationId: getExclusionRule
      summary: Get an exclusion rule (its ETag is required before PATCH — the list endpoint carries no per-item ETag)
      responses:
        '200':
          {
            description: 'OK.',
            content:
              { application/json: { schema: { $ref: '#/components/schemas/ExclusionRule' } } },
          }
        '404': { $ref: '#/components/responses/NotFound' }
    patch:
      tags: [ExclusionRules]
      operationId: updateExclusionRule
      summary: Deactivate a rule
      parameters: [{ $ref: '#/components/parameters/IfMatch' }]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties: { isActive: { type: boolean } }
      responses:
        '200':
          {
            description: 'OK.',
            content:
              { application/json: { schema: { $ref: '#/components/schemas/ExclusionRule' } } },
          }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/StaleResource' }

  # ------------------------------------------------------------- ScanProfiles
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: xenitex_session
      description: >
        Opaque server-side session (SEC-08). State-changing requests also
        require the double-submit CSRF token in the `X-CSRF-Token` header,
        matched against the session's csrfTokenHash server-side.
  parameters:
    Cursor:
      name: cursor
      in: query
      description: Opaque cursor from a previous page's `nextCursor`. Never a raw offset (ADR 0006).
      schema: { type: string }
    Limit:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, maximum: 500, default: 50 }
    IfMatch:
      name: If-Match
      in: header
      required: true
      description: Current ETag of the resource. A stale value returns 409 (ADR 0006).
      schema: { type: string }
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: Client-generated key. A retried request with the same key returns the original response rather than creating a second resource (ADR 0006).
      schema: { type: string }
...
  headers:
...
    CorrelationId:
      description: >-
        OPS-01 — the correlation identifier for this request, propagated from
        the HTTP request through any queue job and scanner invocation it
        triggers. Echoes an inbound `X-Correlation-Id` if the client sent
        one, otherwise server-generated. Surfaced on error responses so a UI
        can quote it back in a support conversation (see ErrorState).
      schema: { type: string }
  responses:
...
    ValidationError:
      description: Schema validation failed at the API boundary (SEC-12) — unknown fields are rejected, not ignored.
      headers: { X-Correlation-Id: { $ref: '#/components/headers/CorrelationId' } }
      content:
        { application/problem+json: { schema: { $ref: '#/components/schemas/ProblemDetails' } } }
...
    NotFound:
      description: No such resource.
      headers: { X-Correlation-Id: { $ref: '#/components/headers/CorrelationId' } }
      content:
        { application/problem+json: { schema: { $ref: '#/components/schemas/ProblemDetails' } } }
...
    StaleResource:
      description: 'code: concurrency.stale_resource — the supplied If-Match no longer matches (ADR 0006).'
      headers: { X-Correlation-Id: { $ref: '#/components/headers/CorrelationId' } }
      content:
        { application/problem+json: { schema: { $ref: '#/components/schemas/ProblemDetails' } } }
...
  schemas:
    ProblemDetails:
      type: object
      required: [type, title, status, code]
      properties:
        type: { type: string, format: uri }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        instance: { type: string }
        code:
          type: string
          description: Stable, machine-readable, namespaced (e.g. `scope.exclusion_violation`). Clients branch on this, never on title/detail (ADR 0006).
...
    AttestationType:
      { type: string, enum: [self_attested_owner, delegated_authority, contract_engagement, other] }
    ExclusionRuleType: { type: string, enum: [address, range, port, tag] }
...
    AuthorizedScopeCreate:
      type: object
      required: [name, attestationType, attestationDetails]
      properties:
        name: { type: string }
        cidrRanges: { type: array, items: { type: string } }
        hostnames: { type: array, items: { type: string } }
        attestationType: { $ref: '#/components/schemas/AttestationType' }
        attestationDetails: { type: string }
    AuthorizedScope:
      allOf:
        - $ref: '#/components/schemas/AuthorizedScopeCreate'
        - type: object
          required: [id, acceptedByUserId, acceptedAt]
          properties:
            id: { type: string }
            acceptedByUserId: { type: string }
            acceptedAt: { type: string, format: date-time }
            supersededById: { oneOf: [{ type: string }, { type: 'null' }] }
    ExclusionRuleCreate:
      type: object
      required: [ruleType, value, reason]
      properties:
        scopeId:
          { oneOf: [{ type: string }, { type: 'null' }], description: 'null = applies globally' }
        ruleType: { $ref: '#/components/schemas/ExclusionRuleType' }
        value: { type: string }
        reason: { type: string }
    ExclusionRule:
      allOf:
        - $ref: '#/components/schemas/ExclusionRuleCreate'
        - type: object
          required: [id, createdBy, isActive]
          properties:
            id: { type: string }
            createdBy: { type: string }
            isActive: { type: boolean }
...